requirements.txt
rumps
pypinyin
jieba
pyyaml 